refactor: move recommendation logic to component
